almost done

+++ src/scenes/resultWizard.ts
@@ -2,6 +2,11 @@ import { Markup, Scenes } from "telegraf";
 import ResultContext from "../types/resultContext";
 import fetchCourses from "../services/fetchCourses";
 import fetchPublishedResults from "../services/fetchPublishedResults";
+import fetchResult from "../services/fetchResult";
+import formatDob from "../utils/formatDob";
+import formatResultMessage from "../utils/formatResultMessage";
+import formatSummaryMessage from "../utils/formatSummaryMessage";
+import calculateSgpa from "../utils/calculateSgpa";
 
 const resultWizard = new Scenes.WizardScene<ResultContext>(
   "result-wizard",
@@ -12,23 +17,24 @@ const resultWizard = new Scenes.WizardScene<ResultContext>(
         text: name,
         callback_data: `course_${id}`,
       }));
-      const keyboard = Markup.inlineKeyboard(courseButtons, { columns: 2 });
-      ctx.sendMessage("Choose a course:", keyboard);
-      ctx.deleteMessage();
+      const keyboard = Markup.inlineKeyboard(courseButtons, { columns: 1 });
+      await ctx.sendMessage("Choose a course:", keyboard);
       return ctx.wizard.next();
     } catch (error) {
-      await ctx.reply("Something went wrong. Please try again later.");
-      return await ctx.scene.leave();
+      return handleError(ctx, error);
     }
   },
   async (ctx) => {
+    if (ctx.message) {
+      return ctx.reply("Please choose a valid option");
+    }
+    await ctx.deleteMessage();
     const courseId: number = Number.parseInt(
       (ctx.callbackQuery as any)?.data?.split("_")[1],
     );
     ctx.session.courseId = courseId;
     try {
       const publishedResults = await fetchPublishedResults(courseId);
-      console.log(publishedResults);
       const resultButtons = publishedResults.map(
         ({ resultName, examDefId, schemeId }) => ({
           text: resultName,
@@ -36,36 +42,62 @@ const resultWizard = new Scenes.WizardScene<ResultContext>(
         }),
       );
       const keyboard = Markup.inlineKeyboard(resultButtons, { columns: 1 });
-      ctx.sendMessage("Choose a result:", keyboard);
-      ctx.deleteMessage();
+      await ctx.sendMessage("Choose a result:", keyboard);
       return ctx.wizard.next();
     } catch (error) {
-      console.error(error);
-      ctx.reply("Something went wrong. Please try again later.");
-      return await ctx.scene.leave();
+      return handleError(ctx, error);
     }
   },
   async (ctx) => {
     const [examDefId, schemeId] = (ctx.callbackQuery as any)?.data?.split("_");
-    console.log(examDefId, schemeId);
+    await ctx.deleteMessage();
+    ctx.session.examDefId = examDefId;
+    ctx.session.schemeId = schemeId;
+    await ctx.reply("Please enter your KTU Registration Number");
     return ctx.wizard.next();
   },
   async (ctx) => {
-    const ktuid: string = (ctx.message as any)?.text || "";
-    ctx.session.ktuid = ktuid;
-    await ctx.reply("Enter your fucking dob");
+    const regisNo: string = (ctx.message as any)?.text;
+    if (!regisNo) {
+      return ctx.reply("Please enter a valid registration number");
+    }
+    ctx.session.regisNo = regisNo.toUpperCase();
+    await ctx.reply("Enter your Date of Birth (DD/MM/YYYY)");
     return ctx.wizard.next();
   },
   async (ctx) => {
-    const dob: string = (ctx.message as any)?.text || "";
+    let dob: string = (ctx.message as any)?.text;
+    if (!dob) {
+      return ctx.reply("Please enter a valid date of birth");
+    }
+    try {
+      dob = formatDob(dob);
+    } catch (error) {
+      return ctx.reply("Please enter a valid date of birth");
+    }
     ctx.session.dob = dob;
-    return ctx.wizard.next();
-  },
-  async (ctx) => {
-    await ctx.reply("Done");
-    await ctx.reply(ctx.session.course);
-    return await ctx.scene.leave();
+    try {
+      const { summary, resultDetails } = await fetchResult(
+        ctx.session.dob,
+        ctx.session.regisNo,
+        ctx.session.examDefId,
+        ctx.session.schemeId,
+      );
+
+      const sgpa = calculateSgpa(resultDetails);
+      await ctx.replyWithHTML(formatSummaryMessage(summary));
+      await ctx.replyWithHTML(formatResultMessage(resultDetails, sgpa));
+      return ctx.scene.leave();
+    } catch (error) {
+      return handleError(ctx, error);
+    }
   },
 );
+
+function handleError(ctx: ResultContext, error: any) {
+  console.error(error);
+  ctx.reply("Something went wrong. Please try again later.");
+  return ctx.scene.leave();
+}
 
 export default resultWizard;
